+++ targets/opsis/hdmi2usb.py
@@ -5,26 +5,20 @@ from litex.soc.interconnect import stream
 from gateware.encoder import EncoderDMAReader, EncoderBuffer, Encoder
 from gateware.streamer import USBStreamer
 
-from targets.utils import csr_map_update
 from targets.opsis.video import SoC as BaseSoC
 
 
 class HDMI2USBSoC(BaseSoC):
-    csr_peripherals = (
-        "encoder_reader",
-        "encoder",
-    )
-    csr_map_update(BaseSoC.csr_map, csr_peripherals)
-    mem_map = {
-        "encoder": 0xd0000000
-    }
-    mem_map.update(BaseSoC.mem_map)
+    mem_map = {**BaseSoC.mem_map, **{
+        "encoder": 0xd0000000,
+    }}
 
     def __init__(self, platform, *args, **kwargs):
         BaseSoC.__init__(self, platform, *args, **kwargs)
 
         encoder_port = self.sdram.crossbar.get_port()
         self.submodules.encoder_reader = EncoderDMAReader(encoder_port)
+        self.add_csr("encoder_reader")
         encoder_cdc = stream.AsyncFIFO([("data", 128)], 4)
         encoder_cdc = ClockDomainsRenamer({"write": "sys",
                                            "read": "encoder"})(encoder_cdc)
@@ -32,6 +26,7 @@ class HDMI2USBSoC(BaseSoC):
         encoder = Encoder(platform)
         encoder_streamer = USBStreamer(platform, platform.request("fx2"))
         self.submodules += encoder_cdc, encoder_buffer, encoder, encoder_streamer
+        self.add_csr("encoder")
 
         self.comb += [
             self.encoder_reader.source.connect(encoder_cdc.sink),
